+++ packages/api/bin/generate-environment-md.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+yarn ts-node ./bin/generate-environment-md.mts > ENVIRONMENT.md
+yarn prettier --write ENVIRONMENT.md