+++ packages/api/src/business/auth/objects/LoginFailureResponse.ts
@@ -1,0 +1,7 @@
+import { Field, ObjectType } from "type-graphql";
+
+@ObjectType()
+export class LoginFailureResponse {
+  @Field(() => String)
+  message: string = "Invalid username or password";
+}