+++ packages/api/src/errors/NotFoundError.ts
@@ -3,7 +3,7 @@ import { BaseError } from "./BaseError.js";
 
 @ObjectType({ implements: BaseError })
 export class NotFoundError extends BaseError {
-  constructor() {
-    super("Entity not found");
+  constructor(message = "Entity not found") {
+    super(message);
   }
 }