+++ packages/api/src/ses/sesConfig.ts
@@ -1,0 +1,9 @@
+import { SESClientConfig } from "@aws-sdk/client-ses";
+import { AppGraphqlContext } from "../graphql/AppGraphqlContext.js";
+
+export const sesConfig = ({
+  sesEnvironment,
+}: AppGraphqlContext): SESClientConfig => ({
+  endpoint: sesEnvironment.endpoint || undefined,
+  region: "us-east-1",
+});