+++ packages/client/src/ui/AppRoutes.ts
@@ -4,6 +4,8 @@ import { uuidToSlug } from "../utils/uuidUtils";
 export const AppRouteParts = {
   admin: () => ["admin"],
   inviteCodeList: () => [...AppRouteParts.admin(), "inviteCodes"],
+  forgotPassword: () => ["forgot-password"],
+  resetPassword: () => ["reset-password"],
   login: () => ["login"],
   logout: () => ["logout"],
   register: () => ["register"],